Refactor meta box to controller
